requirements.txt
fastapi>=0.95.2
uvicorn>=0.23.0
pydantic>=2.7.3
sqlalchemy>=2.0.17
passlib>=1.7.4
python-jose>=3.3.0
email-validator>=1.3.1
requests>=2.31.0
python-dotenv>=1.0.0
bcrypt>=4.0.0
aiohttp>=3.8.5
psycopg2>=2.9.8
twilio>=9.3.0
python-multipart>=0.0.17
jwt>=1.3.0
uvicorn>=0.32.0
pydantic
Pylint
sendgrid>=6.11.0
mailersend~=0.5.8
psycopg2
asyncpg
pytest~=8.3.3
httpx
oci-cli~=3.50.2
alembic~=1.14.0
starlette~=0.41.3
PyJWT~=2.10.1